Change HID_::SendReport "id" argument from 16 to 8bits

+++ src/HIDPowerDevice.h
@@ -113,8 +113,7 @@ private:
 public:
   HIDPowerDevice_(void);
   
-
-  int sendDate(uint16_t id, uint16_t year, uint8_t month, uint8_t day);
+  int sendDate(uint8_t id, uint16_t year, uint8_t month, uint8_t day);
   
   int setStringFeature(uint8_t id, const uint8_t* index, const char* data);
 };
